\name{Topotecan}
\alias{Topotecan}
\docType{data}

\title{Example dataset} 

\description{ A subset drug response data extracted from the cancer cell line encyclopedia (CCLE) database for the drug topotecan. 
}
\usage{data(Topotecan)}
\format{
\describe{
  \item{ X }{ a \emph{n}x\emph{p} data matrix; n=491, p=89}
  \item{ Y  }{ response vector.}
}
}


\references{
Liang, F., Li, Q. and Zhou, L. (2017) Bayesian Neural Networks for Selection of Drug Sensitive Genes. Journal of the American Statistical Association, under revision.

Xue, J. and Liang, F. (2017). Robust model-free feature screening for ultrahigh dimensional data. Journal of Computational and Graphical Statistics, in press. 
}


\keyword{datasets}
